Add datepicker for scheduling gift cards (#2540)
* Add datepicker for scheduling gift cards

We're using a native date picker for now as this was deemed to be the best starting point int terms of utility and accessibility. As the feature matures we may choose to implement something custom which is more themable but this is a good starting position.

Other theme developers may use a different date picker as long as the line itmem property is in ISO format YYYY-MM-DD

We also now use error messages which contain the field as part of the error string which is much better for localisation. Unfortunately it would be a breaking change to introduce this, so we've made it opt in my keying on the new _shopify_offset parameter. It's a bit strange to use this property as the switch but it works.

We want the TZ offset so that we can schedule the gift card send for a sensible time in the user's local timezone rather than in the shop timezone which might cause confusion.

Whilst I was here I also cleaned up a minor bug with empty line item properties being sent even when the checkbox was not checked. I've done this simply by disabling the fields whenever the checkbox is unchecked.

* Rename recipient fields for consistency

* invert if/else for readibilty

* Update 20 translation files

* Update 10 translation files

---------

Co-authored-by: translation-platform[bot] <34770790+translation-platform[bot]@users.noreply.github.com>